Improve admin interface
Next I want to be able to create an admin changelist view for
wordsets. I don't mind creating a model for wordsets, but ideally the
changelis view would be displayed on the fly, without committing the
wordsets to the db.

Similarly, I want to be able to display a changelist view for
arbitrary subsets of items (e.g. those that contain a certain
wordset).